+++ flmodules/src/dht_storage/mod.rs
@@ -1,7 +1,7 @@
 // The core algorithm of this module
 pub mod core;
 // Messages for this module
-pub mod messages;
+pub mod intern;
 // Integrating with other modules
 pub mod broker;
 // Easy interaction with dht_storage and realms