+++ flmodules/src/gossip_events/intern.rs
@@ -7,6 +7,12 @@ use flarch::{
 use serde::{Deserialize, Serialize};
 use tokio::sync::watch;
 
+use crate::{
+    gossip_events::broker::{GossipIn, GossipOut},
+    random_connections::broker::{RandomIn, RandomOut},
+    router::messages::NetworkWrapper,
+};
+
 use super::{broker::MODULE_NAME, core::*};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -17,25 +23,24 @@ pub enum ModuleMessage {
     RequestEvents(Vec<U256>),
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum GossipIn {
+#[derive(Debug, Clone, PartialEq)]
+pub(super) enum InternIn {
     Tick,
-    FromNetwork(NodeID, ModuleMessage),
-    AddEvent(Event),
-    UpdateNodeList(NodeIDs),
+    Network(RandomOut),
+    Gossip(GossipIn),
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum GossipOut {
-    ToNetwork(NodeID, ModuleMessage),
-    NewEvent(Event),
+#[derive(Debug, Clone, PartialEq)]
+pub(super) enum InternOut {
+    Network(RandomIn),
+    Gossip(GossipOut),
 }
 
 /// The first module to use the random_connections is a copy of the previous
 /// chat.
 /// Now it holds events of multiple categories and exchanges them between the
 /// nodes.
-pub struct Messages {
+pub(super) struct Intern {
     events: EventsStorage,
     storage: Box<dyn DataStorage + Send>,
     tx: Option<watch::Sender<EventsStorage>>,
@@ -45,7 +50,7 @@ pub struct Messages {
     ticks: usize,
 }
 
-impl std::fmt::Debug for Messages {
+impl std::fmt::Debug for Intern {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("GossipEvents")
             .field("events", &self.events)
@@ -56,7 +61,7 @@ impl std::fmt::Debug for Messages {
     }
 }
 
-impl Messages {
+impl Intern {
     /// Returns a new chat module.
     pub fn new(
         id: NodeID,
@@ -97,37 +102,36 @@ impl Messages {
             .map(|e| log::warn!("Couldn't store gossip: {e:?}"));
     }
 
-    /// Processes many messages at a time.
-    pub fn process_messages(&mut self, msgs: Vec<GossipIn>) -> anyhow::Result<Vec<GossipOut>> {
-        let mut out = vec![];
-        for msg in msgs {
-            out.extend(self.process_message(msg));
+    fn msg_tick(&mut self) -> Vec<InternOut> {
+        self.ticks += 1;
+        // Magic number - 3 seems to work fine here...
+        if self.ticks >= 3 {
+            self.tick()
+        } else {
+            vec![]
         }
-        Ok(out)
     }
 
-    /// Processes one generic message and returns the result.
-    pub fn process_message(&mut self, msg: GossipIn) -> Vec<GossipOut> {
-        // log::trace!("{} got message {:?}", self.id, msg);
+    fn msg_net(&mut self, msg: RandomOut) -> Vec<InternOut> {
         match msg {
-            GossipIn::Tick => {
-                self.ticks += 1;
-                // Magic number - 3 seems to work fine here...
-                if self.ticks >= 3 {
-                    self.tick()
-                } else {
-                    vec![]
-                }
-            }
-            GossipIn::FromNetwork(src, node_msg) => self.process_node_message(src, node_msg),
-            GossipIn::AddEvent(ev) => self.add_event(ev),
-            GossipIn::UpdateNodeList(ids) => self.node_list(ids),
+            RandomOut::NodeIDsConnected(node_ids) => self.node_list(node_ids),
+            RandomOut::NetworkWrapperFromNetwork(src, network_wrapper) => network_wrapper
+                .unwrap_yaml(MODULE_NAME)
+                .map(|msg| self.process_node_message(src, msg))
+                .unwrap_or(vec![]),
+            _ => vec![],
+        }
+    }
+
+    fn msg_gossip(&mut self, msg: GossipIn) -> Vec<InternOut> {
+        match msg {
+            GossipIn::AddEvent(event) => self.add_event(event),
         }
     }
 
     /// Processes a node to node message and returns zero or more
     /// MessageOut.
-    pub fn process_node_message(&mut self, src: NodeID, msg: ModuleMessage) -> Vec<GossipOut> {
+    fn process_node_message(&mut self, src: NodeID, msg: ModuleMessage) -> Vec<InternOut> {
         match msg {
             ModuleMessage::KnownEventIDs(ids) => self.node_known_event_ids(src, ids),
             ModuleMessage::Events(events) => self.node_events(src, events),
@@ -138,7 +142,7 @@ impl Messages {
 
     /// Adds an event if it's not known yet or not too old.
     /// This will send out the event to all other nodes.
-    pub fn add_event(&mut self, event: Event) -> Vec<GossipOut> {
+    fn add_event(&mut self, event: Event) -> Vec<InternOut> {
         if self.events.add_event(event.clone()) {
             self.store();
             return self.send_events(self.id, &[event]);
@@ -148,7 +152,7 @@ impl Messages {
 
     /// Takes a vector of events and stores the new events. It returns all
     /// events that are new to the system.
-    pub fn add_events(&mut self, events: Vec<Event>) -> Vec<Event> {
+    fn add_events(&mut self, events: Vec<Event>) -> Vec<Event> {
         events
             .into_iter()
             .inspect(|e| self.outstanding.retain(|os| os != &e.get_id()))
@@ -156,13 +160,13 @@ impl Messages {
             .collect()
     }
 
-    fn send_events(&self, src: NodeID, events: &[Event]) -> Vec<GossipOut> {
+    fn send_events(&self, src: NodeID, events: &[Event]) -> Vec<InternOut> {
         self.nodes
             .0
             .iter()
             .filter(|&&node_id| node_id != src && node_id != self.id)
-            .map(|node_id| {
-                GossipOut::ToNetwork(
+            .flat_map(|node_id| {
+                Self::create_network_msg(
                     *node_id,
                     ModuleMessage::KnownEventIDs(
                         events
@@ -178,18 +182,22 @@ impl Messages {
                     ),
                 )
             })
-            .chain(events.iter().map(|e| GossipOut::NewEvent(e.clone())))
+            .chain(
+                events
+                    .iter()
+                    .map(|e| InternOut::Gossip(GossipOut::NewEvent(e.clone()))),
+            )
             .collect()
     }
 
     /// If an updated list of nodes is available, send a `RequestEventIDs` to
     /// all new nodes.
-    pub fn node_list(&mut self, ids: NodeIDs) -> Vec<GossipOut> {
+    fn node_list(&mut self, ids: NodeIDs) -> Vec<InternOut> {
         let reply = ids
             .0
             .iter()
             .filter(|&id| !self.nodes.0.contains(id) && id != &self.id)
-            .map(|&id| GossipOut::ToNetwork(id, ModuleMessage::RequestEventIDs))
+            .flat_map(|&id| Self::create_network_msg(id, ModuleMessage::RequestEventIDs))
             .collect();
         self.nodes = ids;
         reply
@@ -198,25 +206,22 @@ impl Messages {
     /// Reply with a list of events this node doesn't know yet.
     /// We suppose that if there are too old events in here, they will be
     /// discarded over time.
-    pub fn node_known_event_ids(&mut self, src: NodeID, ids: Vec<U256>) -> Vec<GossipOut> {
+    fn node_known_event_ids(&mut self, src: NodeID, ids: Vec<U256>) -> Vec<InternOut> {
         let unknown_ids = self.filter_known_events(ids);
         if !unknown_ids.is_empty() {
             self.outstanding.extend(unknown_ids.clone());
-            return vec![GossipOut::ToNetwork(
-                src,
-                ModuleMessage::RequestEvents(unknown_ids),
-            )];
+            return Self::create_network_msg(src, ModuleMessage::RequestEvents(unknown_ids));
         }
         vec![]
     }
 
     /// Store the new events and send them to the other nodes.
-    pub fn node_events(&mut self, src: NodeID, events: Vec<Event>) -> Vec<GossipOut> {
+    fn node_events(&mut self, src: NodeID, events: Vec<Event>) -> Vec<InternOut> {
         // Attention: self.send_event can return an empty vec in case there are no
         // other nodes available yet. So it's not enough to check the 'output' variable
         // to know if the MessageOut::Updated needs to be sent or not.
         let events_out = self.add_events(events);
-        let output: Vec<GossipOut> = self.send_events(src, &events_out);
+        let output: Vec<InternOut> = self.send_events(src, &events_out);
         if !events_out.is_empty() {
             self.store();
         }
@@ -225,25 +230,22 @@ impl Messages {
 
     /// Send the events to the other node. One or more of the requested
     /// events might be missing.
-    pub fn node_request_events(&mut self, src: NodeID, ids: Vec<U256>) -> Vec<GossipOut> {
+    fn node_request_events(&mut self, src: NodeID, ids: Vec<U256>) -> Vec<InternOut> {
         let events: Vec<Event> = self.events.get_events_by_ids(ids);
         if !events.is_empty() {
-            vec![GossipOut::ToNetwork(src, ModuleMessage::Events(events))]
+            Self::create_network_msg(src, ModuleMessage::Events(events))
         } else {
             vec![]
         }
     }
 
     /// Returns the list of known events.
-    pub fn node_request_event_list(&mut self, src: NodeID) -> Vec<GossipOut> {
-        vec![GossipOut::ToNetwork(
-            src,
-            ModuleMessage::KnownEventIDs(self.events.event_ids()),
-        )]
+    fn node_request_event_list(&mut self, src: NodeID) -> Vec<InternOut> {
+        Self::create_network_msg(src, ModuleMessage::KnownEventIDs(self.events.event_ids()))
     }
 
     /// Returns all ids that are not in our storage
-    pub fn filter_known_events(&self, eventids: Vec<U256>) -> Vec<U256> {
+    fn filter_known_events(&self, eventids: Vec<U256>) -> Vec<U256> {
         let our_ids = self.events.event_ids();
         eventids
             .into_iter()
@@ -251,26 +253,36 @@ impl Messages {
             .collect()
     }
 
-    pub fn storage(&self) -> EventsStorage {
-        self.events.clone()
-    }
-
     /// Every tick clear the outstanding vector and request new event IDs.
-    pub fn tick(&mut self) -> Vec<GossipOut> {
+    fn tick(&mut self) -> Vec<InternOut> {
         self.outstanding.clear();
         self.nodes
             .0
             .iter()
-            .map(|id| GossipOut::ToNetwork(*id, ModuleMessage::RequestEventIDs))
+            .flat_map(|id| Self::create_network_msg(*id, ModuleMessage::RequestEventIDs))
             .collect()
+    }
+
+    fn create_network_msg(dst: NodeID, msg: ModuleMessage) -> Vec<InternOut> {
+        NetworkWrapper::wrap_yaml(MODULE_NAME, &msg)
+            .map(|net_msg| {
+                vec![InternOut::Network(RandomIn::NetworkWrapperToNetwork(
+                    dst, net_msg,
+                ))]
+            })
+            .unwrap_or(vec![])
     }
 }
 
 #[platform_async_trait()]
-impl SubsystemHandler<GossipIn, GossipOut> for Messages {
-    async fn messages(&mut self, msgs: Vec<GossipIn>) -> Vec<GossipOut> {
+impl SubsystemHandler<InternIn, InternOut> for Intern {
+    async fn messages(&mut self, msgs: Vec<InternIn>) -> Vec<InternOut> {
         msgs.into_iter()
-            .flat_map(|msg| self.process_message(msg))
+            .flat_map(|msg| match msg {
+                InternIn::Tick => self.msg_tick(),
+                InternIn::Network(random_out) => self.msg_net(random_out),
+                InternIn::Gossip(gossip_in) => self.msg_gossip(gossip_in),
+            })
             .collect()
     }
 }