+++ flmodules/src/template/mod.rs
@@ -1,6 +1,3 @@
-// The core algorithm of this module
-pub mod core;
-// Messages for this module
-pub mod messages;
-// Integrating with other modules
-pub mod broker;
+pub mod multi;
+pub mod persistent;
+pub mod simple;