+++ flmodules/src/template/multi/mod.rs
@@ -1,3 +1,2 @@
 pub mod broker;
-pub mod core;
 pub mod intern;